requirements.txt
numpy
pillow
torch
torchaudio
torchvision
tqdm
albumentations
